Refactor: abstracted logic to a custom hook on view charts
